refactor(scalars): move PHIDInput to ui
