+++ packages/design-system/src/scalars/components/phid-field/index.ts
@@ -1,1 +1,2 @@
-export * from "./phid-field.js";
+export { PHIDField } from "./phid-field.js";
+export type { PHIDFieldProps } from "./types.js";
