+++ packages/design-system/src/scalars/components/phid-field/phid-field.test.tsx
@@ -1,6 +1,7 @@
 import { Form } from "#scalars";
 import { render, screen, waitFor } from "@testing-library/react";
 import { userEvent } from "@testing-library/user-event";
+import { describe, expect, it, vi } from "vitest";
 import { renderWithForm } from "../../lib/testing.js";
 import { PHIDField } from "./phid-field.js";
