+++ packages/design-system/src/ui/components/data-entry/phid-input/index.ts
@@ -1,0 +1,2 @@
+export { PHIDInput } from "./phid-input.js";
+export type { PHIDInputProps, PHIDOption } from "./types.js";
